+++ src/Tests/Network.fs
@@ -1212,4 +1212,113 @@ let tests =
                 "[resourceId(\u0027Microsoft.Network/virtualNetworks/subnets\u0027, \u0027test-vnet\u0027, \u0027test-subnet\u0027)]"
                 "Incorrect subnet id for ipConfig"
         }
+
+        test "Automatically carved subnets with route table" {
+            let myRt = routeTable { name "my-rt" }
+
+            let vnetName = "my-vnet"
+            let webSubnet = "web"
+            let appsSubnet = "apps"
+            let noRtSubnet = "no-rt"
+
+            let myNet = vnet {
+                name vnetName
+
+                build_address_spaces [
+                    addressSpace {
+                        space "10.28.0.0/16"
+
+                        subnets [
+                            subnetSpec {
+                                name webSubnet
+                                size 24
+                                route_table myRt
+                            }
+                            subnetSpec {
+                                name appsSubnet
+                                size 24
+                                route_table myRt
+                            }
+                            subnetSpec {
+                                name noRtSubnet
+                                size 24
+                            }
+                        ]
+                    }
+                ]
+            }
+
+            let template = arm { add_resources [ myNet; myRt ] }
+            let jobj = template.Template |> Writer.toJson |> Newtonsoft.Json.Linq.JObject.Parse
+
+            let dependencies =
+                jobj.SelectToken "resources[?(@.type=='Microsoft.Network/virtualNetworks')].dependsOn"
+                :?> Newtonsoft.Json.Linq.JArray
+
+            Expect.isNotNull dependencies "vnet missing dependency for rt"
+            Expect.hasLength dependencies 1 "Incorrect number of dependencies for vnet"
+
+            Expect.equal
+                (dependencies.[0].ToString())
+                "[resourceId('Microsoft.Network/routeTables', 'my-rt')]"
+                "Incorrect vnet dependencies"
+
+            let vnet = template |> getVnetResource
+            Expect.isNotNull vnet.Subnets.[0].RouteTable "First subnet missing route table"
+
+            Expect.equal
+                vnet.Subnets.[0].RouteTable.Id
+                "[resourceId('Microsoft.Network/routeTables', 'my-rt')]"
+                "Incorrect route table for first subnet"
+
+            Expect.isNotNull vnet.Subnets.[0].RouteTable "Second subnet missing route table"
+
+            Expect.equal
+                vnet.Subnets.[1].RouteTable.Id
+                "[resourceId('Microsoft.Network/routeTables', 'my-rt')]"
+                "Incorrect route table for second subnet"
+
+            Expect.isNull vnet.Subnets.[2].RouteTable "Third subnet should not have route table"
+        }
+
+        test "Vnet with linked route table doesn't add dependsOn" {
+            let vnetName = "my-vnet"
+            let webSubnet = "web"
+
+            let myNet = vnet {
+                name vnetName
+
+                build_address_spaces [
+                    addressSpace {
+                        space "10.28.0.0/16"
+
+                        subnets [
+                            subnetSpec {
+                                name webSubnet
+                                size 24
+
+                                link_to_route_table (routeTables.resourceId "my-rt")
+                            }
+                        ]
+                    }
+                ]
+            }
+
+            let template = arm { add_resources [ myNet ] }
+            let jobj = template.Template |> Writer.toJson |> Newtonsoft.Json.Linq.JObject.Parse
+
+            let dependencies =
+                jobj.SelectToken "resources[?(@.type=='Microsoft.Network/virtualNetworks')].dependsOn"
+                :?> Newtonsoft.Json.Linq.JArray
+
+            Expect.hasLength dependencies 0 "Should be no vnet dependencies when linking to route table"
+
+            let vnet = template |> getVnetResource
+            Expect.isNotNull vnet.Subnets.[0].RouteTable "Subnet missing route table"
+
+            Expect.equal
+                vnet.Subnets.[0].RouteTable.Id
+                "[resourceId('Microsoft.Network/routeTables', 'my-rt')]"
+                "Incorrect route table for subnet"
+        }
     ]