+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ Release Notes
 
 ## 1.9.4
 * Network Security Groups: Use a common protocol in security rules with multiple sources. Defaults to Any if sources use different protocols.
+* Virtual Network: Specify the route table for a subnet.
 
 ## 1.9.3
 * Deployments: Default to resource group location rather than West Europe.
@@ -13,6 +14,7 @@ Release Notes
 * Functions: Support automatic creation of Workspace Based App Insights & Log Analytics.
 * Az: Update `ad` commands to work with latest (breaking) structure.
 * PostgreSQL: Fix a number of issues around the introduction of Flexible Servers.
+
 
 ## 1.9.2
 * Container Apps: Fix to container registry credential to not emit a secret for a managed identity.